Add possibility to use 'dev' mode to avoid needing database during development.
