+++ API/package.json
@@ -1,4 +1,7 @@
 {
+  "scripts": {
+    "start": "tsc && node build/server.js"
+  },
   "dependencies": {
     "@types/body-parser": "^1.19.5",
     "@types/cors": "^2.8.17",
