+++ API/src/routes/controllers.ts
@@ -9,11 +9,12 @@ import {
 
 const controllersRoute = Router();
 
+// Returns all of the available controllers.
 controllersRoute.get("/controllers", (req, res) => {
     getControllers(req, res);
 });
 
-controllersRoute.post("/controller", (req, res) => {
+controllersRoute.post("/controllers", (req, res) => {
     addController(req, res);
 });
 
@@ -21,6 +22,7 @@ controllersRoute.put("/controller/:id", (req, res) => {
     editController(req, res);
 });
 
+// Get a specific controller by CID.
 controllersRoute.get("/controller/:id", (req, res) => {
     getController(req, res);
 });
